+++ lib/src/flag_manager.dart
@@ -1,7 +1,8 @@
 import 'package:file_state_manager/file_state_manager.dart';
 
 ///
-/// This package manages multiple flags identified by name and supports serialization and deserialization.
+/// This class manages multiple flags identified by name and
+/// supports serialization and deserialization.
 ///
 /// Author Masahide Mori
 ///
@@ -9,7 +10,7 @@ import 'package:file_state_manager/file_state_manager.dart';
 ///
 class FlagManager extends CloneableFile {
   static const String className = 'FlagManager';
-  static const String version = '4';
+  static const String version = '5';
   final Map<String, bool> _map = {};
   static const String _saveKey = 'map';
 
@@ -60,12 +61,11 @@ class FlagManager extends CloneableFile {
     return d;
   }
 
-  /// (en)Returns an flag with the specified name if it has been generated,
-  /// otherwise it is generated.
+  /// (en)Returns the value corresponding to the given name.
   ///
-  /// (ja)指定名のフラグが生成済みならばそれを、無ければ生成して返します。
+  /// (ja)指定名に対応する値を返します。
   ///
-  /// * [name] : A unique name assigned to the flag.
+  /// * [name] : Target name.
   /// * [initialValue] : Initial value. Applies only when first created.
   /// * [isAlwaysInitialize] : If true, always set initialValue.
   bool getFlag(String name,
@@ -81,10 +81,12 @@ class FlagManager extends CloneableFile {
     }
   }
 
-  /// (en) If an variable with the specified name has already been created,
-  /// value is assigned to it, otherwise value is assigned to a newly created variable.
+  /// (en) Overwrites the value managed by this class that
+  /// corresponds to the specified name.
+  /// If it does not exist, it will be added.
   ///
-  /// (ja)指定名の変数が生成済みならばそれに代入し、無ければ新規生成した変数に代入します。
+  /// (ja)このクラスで管理中の、指定の名前に対応する値を上書きします。
+  /// 存在しない場合は管理対象が追加されます。
   void setFlag(String name, bool value) {
     _map[name] = value;
   }