+++ lib/src/value_manager.dart
@@ -1,7 +1,8 @@
 import 'package:file_state_manager/file_state_manager.dart';
 
 ///
-/// This package manages multiple values identified by name and supports serialization and deserialization.
+/// This class manages multiple values identified by name and
+/// supports serialization and deserialization.
 ///
 /// Author Masahide Mori
 ///
@@ -9,7 +10,7 @@ import 'package:file_state_manager/file_state_manager.dart';
 ///
 class ValueManager extends CloneableFile {
   static const String className = 'ValueManager';
-  static const String version = '5';
+  static const String version = '6';
   final Map<String, double?> _map = {};
   static const String _saveKey = 'map';
 
@@ -60,12 +61,11 @@ class ValueManager extends CloneableFile {
     return d;
   }
 
-  /// (en)Returns an value with the specified name if it has been generated,
-  /// otherwise it is generated.
+  /// (en)Returns the value corresponding to the given name.
   ///
-  /// (ja)指定名に対応する数値が生成済みならばそれを、無ければ生成して返します。
+  /// (ja)指定名に対応する値を返します。
   ///
-  /// * [name] : A unique name assigned to the value.
+  /// * [name] : Target name.
   /// * [initialValue] : Initial value. Applies only when first created.
   /// * [isAlwaysInitialize] : If true, always set initialValue.
   double? getValue(String name,
@@ -81,10 +81,12 @@ class ValueManager extends CloneableFile {
     }
   }
 
-  /// (en) If an variable with the specified name has already been created,
-  /// value is assigned to it, otherwise value is assigned to a newly created variable.
+  /// (en) Overwrites the value managed by this class that
+  /// corresponds to the specified name.
+  /// If it does not exist, it will be added.
   ///
-  /// (ja)指定名の変数が生成済みならばそれに代入し、無ければ新規生成した変数に代入します。
+  /// (ja)このクラスで管理中の、指定の名前に対応する値を上書きします。
+  /// 存在しない場合は管理対象が追加されます。
   void setValue(String name, double? value) {
     _map[name] = value;
   }